Start working on Vencord Desktop support
