+++ src/main/patchWin32Updater.ts
