+++ src/main/updater/common.ts
