+++ src/main/utils/constants.ts
@@ -33,3 +33,5 @@ export const ALLOWED_PROTOCOLS = [
     "steam:",
     "spotify:"
 ];
+
+export const IS_VANILLA = /* @__PURE__ */ process.argv.includes("--vanilla");
