+++ src/main/utils/simpleGet.ts
