+++ src/plugins/settings.tsx
@@ -168,6 +168,7 @@ export default definePlugin({
     get additionalInfo() {
         if (IS_DEV) return " (Dev)";
         if (IS_WEB) return " (Web)";
+        if (IS_VENCORD_DESKTOP) return " (Vencord Desktop)";
         if (IS_STANDALONE) return " (Standalone)";
         return "";
     },
